+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 2.0.0
+
+- complete refactoring to make this module universal
+- MIGRATION:
+    - `require('fluent_conv/ftl2js')` should be replaced with `require('fluent_conv/cjs/ftl2js')`
+    - `require('fluent_conv/js2ftl')` should be replaced with `require('fluent_conv/cjs/js2ftl')`
+
 ### 1.1.1
 
 - getVariant name fix for NumberLiteral variants [2](https://github.com/locize/fluent_conv/pull/2)